Attempted fileSize in archiveSize.c

+++ Assignment 2/archivingSize.h
@@ -1,1 +1,1 @@
-int fileSize(FILE * file);
+long fileSize(FILE * file);
